types: eliminate remaining model-layer mypy errors; flip typecheck blocking
Resolve the final 43 `mypy model/ utils/` errors (5 files) and flip the CI
`typecheck` job from advisory to blocking. Zero runtime behavior change.

B+C sites (26 errors) — typed local bind pattern: bind
`doc: fitz.Document = self.doc` (or `= model.doc`) immediately before the
first flagged use, then dereference `doc`. `fitz` is untyped so
`fitz.Document` is `Any`; assigning `Any | None` to it satisfies mypy while
the runtime semantics (including the AttributeError/TypeError raised on a
None doc) stay identical. Applied in pdf_model.py (_restore_doc_from_snapshot,
_roundtrip_live_doc, _atomic_full_save, _restore_page_from_snapshot,
_full_save_to_path, _render_page_gray_array, export_pages), pdf_object_ops.py
(_rewrite_native_image_matrix, _image_xref_digest, _remove_native_image_invocation,
_rotate_native_image_absolute) and pdf_text_edit.py (edit_text).

The originally-planned raising `_require_doc()` helper was REJECTED at design
time: the current no-doc failure modes are heterogeneous (TypeError from
len(None)/None[idx], swallowed AttributeError in _image_xref_digest, and silent
no-ops), so a unified raising guard would change behavior. Guard unification is
deferred to a future behavior-allowed PR.

D sites (17 errors) — annotations + runtime-inert casts:
- original_path setter param narrowed to `str` (getter stays `str | None`)
- best_key/key pre-declared as `tuple[float, ...]` in _resolve_paragraph_candidate
- source_colors widened to `list[tuple[float, ...]]` in _build_multi_style_html
- `run_color = cast(tuple, new_colors[idx])`
- pdf_content_ops.py rotation params int→float (font-size-class pitfall: never
  coerce float→int) + dead initializer `rotation = 0` → `0.0`
- payload dicts annotated `dict[str, Any]` in the object-marker creators
- ocr_tool.py: `_detector`/`_recognizer` typed `Any`; bbox tuple cast to
  `tuple[float, float, float, float]`; adapter local typed `_SuryaAdapter | None`
  with an inert cast at the call site (mypy re-widens across the loop body)

New test_scripts/test_model_no_doc_behavior_pins.py pins the CURRENT crash
behavior on a document-less model (byte-identical refactor evidence); it passes
before and after.

Validation: mypy 43 -> 0 (Success, 34 files); ruff clean; full suite
1579 passed / 21 skipped / 0 failed (1576 baseline + 3 new pins).

+++ model/pdf_content_ops.py
@@ -328,7 +328,7 @@ def discover_native_image_invocations(doc: fitz.Document, page_num: int) -> list
             cm_index = last_cm_by_depth[-1]
             cm_values = _cm_values_from_operands(operators[cm_index].operands) if cm_index is not None else None
             bbox: fitz.Rect | None = None
-            rotation = 0
+            rotation = 0.0
             if cm_values is not None:
                 bbox = _bbox_from_stream_cm(
                     cm_values,
@@ -504,7 +504,7 @@ def serialize_tokens(tokens: list[ContentToken]) -> bytes:
     return b"\n".join(token.raw for token in tokens) + b"\n"
 
 
-def fitz_rect_to_stream_cm(rect: fitz.Rect, page: fitz.Page, rotation: int) -> list[bytes]:
+def fitz_rect_to_stream_cm(rect: fitz.Rect, page: fitz.Page, rotation: float) -> list[bytes]:
     bbox = fitz.Rect(rect)
     width = float(bbox.width)
     height = float(bbox.height)
@@ -532,7 +532,7 @@ def form_rect_to_stream_cm(
     destination_rect: fitz.Rect,
     current_cm_values: tuple[float, float, float, float, float, float],
     current_page_bbox: fitz.Rect,
-    rotation: int,
+    rotation: float,
 ) -> list[bytes] | None:
     """Build a new ``cm`` for an image drawn inside a Form XObject.
